Gradle hack used to put data files in correct location for distribution
The data files used by the program should be stored in the resources
folder, but Gradle seems to ignore them completely when put there.
This is a workaround using another existing Gradle task (part of
the Distribution plugin), but the task which actually should be
used and isn't working (probably I'm not configuring it right) is
processResources.