Show test data on screen
Next step is to just disable the legacy tests for Travis purposes.
They are all deprecated anyway. If the issue is writing to file or memory usage
then those issues will come up again later on when the tests are migrated to JUnit,
and it'll be easy to pinpoint the cause of the problem.